Refactoring, it now compiles again.
But i commented a lot of stuff out

+++ Net.hs
@@ -6,6 +6,7 @@ import Control.Concurrent.MVar
 import Data.Acid
 import Network
 import Control.Concurrent
+import Control.Exception
 
 import Conn
 import ConnSet
@@ -13,10 +14,11 @@ import World
 import Player
 import qualified Counter
 import Data.Time
+import Mud
 
 listeningThread :: Mud -> Int -> IO ()
 listeningThread mud port = void . forkIO $ do
-  s <- listenOn (PortNumber port)
+  s <- listenOn (PortNumber (fromIntegral port))
   idSrc <- Counter.new
   forever $ do
     (h,_,_) <- accept s -- handle, hostname, port
